pyproject.toml
[project]
name = "info-center"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlmodel",
    "SQLAlchemy",
    "psycopg2-binary",
    "python-dotenv",
    "requests",
    "pymongo"
]
